Add new items to the permanent database of seen items *before* truncating the array.
Previously it would truncate the new items array to 6 items before adding them to the permanent database. This meant that next time the feed was checked (and reported something other than “304 Not Modified”), it would view those items which was cut last, as new items.

Sort the unread items before we truncate, so we always return the newest 5 items. Also ensures that the oldest (of at most the newest 5) is first in the ‘unread’ array.

Since we truncate feeds to at most report 5 items, I removed the code that has it skip all new items when the bot is first started.

Also added to_s for the Feed class so we can see what feed URL is causing an exception instead of the object address.